pyproject.toml
[project]
name = "langgraph-react"
version = "0.1.0"
description = ""
authors = [{name = "Eden Marco", email = "emarco177@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black",
    "isort",
    "python-dotenv",
    "langchain-tavily",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "tavily-python>=0.3.0",
    "langchain-community>=0.2.0",
    "packaging>=24.0",
    "langgraph-cli[inmem]>=0.4.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
